requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.1
azure-core==1.26.3
azure-storage-blob==12.15.0
backcall==0.2.0
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-github-storage==1.0.2
django-phonenumber-field==7.0.2
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.4
entrypoints==0.4
executing==1.2.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jupyter_core==5.1.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
phonenumberslite==8.13.4
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pywin32==305
pyzmq==25.0.0
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
twilio==7.16.2
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
whitenoise==6.3.0
